refactor(code): better project structure & fixed issue with coverage
