+++ biome.json
@@ -4,7 +4,7 @@
     "enabled": true
   },
   "files": {
-    "include": ["./index.ts", "./test/**/*.ts"]
+    "include": ["./src/**/*.ts", "./test/**/*.ts"]
   },
   "linter": {
     "enabled": true,
