+++ package.json
@@ -42,7 +42,7 @@
   "license": "MIT",
   "devDependencies": {
     "@biomejs/biome": "1.5.3",
-    "@types/node": "20.11.5",
+    "@types/node": "20.11.13",
     "@vitest/coverage-v8": "1.2.2",
     "lefthook": "1.6.1",
     "stream-equal": "2.0.1",
