+++ tsconfig.build.json
@@ -1,4 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./index.ts"]
+  "include": [
+    "./src/**/*.ts"
+  ]
 }
