+++ tsup.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from 'tsup'
+import {defineConfig} from 'tsup';
 
 export default defineConfig({
   outDir: './dist',
@@ -10,7 +10,7 @@ export default defineConfig({
   }),
   cjsInterop: true,
   entry: {
-    index: './index.ts',
+    index: './src/index.ts',
   },
   sourcemap: true,
   skipNodeModulesBundle: true,
